add date_produced concern in batches and samples
