+++ app/models/concerns/date_produced.rb
@@ -1,0 +1,31 @@
+module DateProduced
+  extend ActiveSupport::Concern
+
+  included do
+    validates_presence_of :date_produced
+    validate :date_produced_is_a_date
+
+    def date_produced_is_a_date
+      return if date_produced.blank?
+      errors.add(:date_produced, "should be a date in #{self.class.date_produced_placeholder}") unless date_produced.is_a?(Time)
+    end
+
+    def date_produced_description
+      if date_produced.is_a?(Time)
+        return date_produced.strftime(I18n.t('date.input_format.pattern'))
+      end
+
+      date_produced
+    end
+  end
+
+  class_methods do
+    def date_produced_placeholder
+      date_format[:placeholder]
+    end
+
+    def date_format
+      { pattern: I18n.t('date.input_format.pattern'), placeholder: I18n.t('date.input_format.placeholder') }
+    end
+  end
+end
