+++ app/models/sample.rb
@@ -4,6 +4,7 @@ class Sample < ActiveRecord::Base
   include SpecimenRole
   include InactivationMethod
   include SiteContained
+  include DateProduced
 
   belongs_to :patient
   belongs_to :encounter
@@ -77,13 +78,5 @@ class Sample < ActiveRecord::Base
 
   def has_entity_id?
     entity_ids.compact.any?
-  end
-
-  def date_produced_description
-    if date_produced.is_a?(Time)
-      return date_produced.strftime(I18n.t('date.input_format.pattern'))
-    end
-
-    date_produced
   end
 end
