+++ app/views/batches/_form.haml
@@ -14,7 +14,7 @@
         .col.pe-3
           = f.label :date_produced, "production date"
         .col
-          = f.text_field :date_produced, placeholder: @batch_form.date_produced_placeholder
+          = f.text_field :date_produced, placeholder: Batch.date_produced_placeholder
 
       .row
         .col.pe-3
